refactor(tests): extract shared TransferIntegrationTestCase base class
Consolidates 4 independently-duplicated createTransferRequest()/
insertTransferRequest() fixture helpers (CarTransferRepositoryIntegrationTest,
TransferRequestTest, CarTransferWorkflowTest, UserDeletionReassignmentTest)
into one TransferIntegrationTestCase base class, so a new NOT NULL column on
car_transfer_requests only needs a defaults-array update in one place instead
of four. All inserts now go through CarTransferRepository::create() instead
of hand-written raw INSERTs, and tearDown() cleanup failures are logged
instead of silently swallowed.

Closes #1283
