+++ docs/releases/RELEASE_NOTES_v2.29.1.md
@@ -101,3 +101,4 @@ Do the following, in order, **separately for each environment** (test, then prod
 - [#1559](https://github.com/elan-registry/registry/issues/1559) — test: relocate tests/regression/ into tests/unit/regression/ tagged #[Group('regression')] — collapsed the standalone directory/testsuite tier, keeping a single source of truth for the regression-test structural check that pre-commit and CI enforce
 - [#1536](https://github.com/elan-registry/registry/issues/1536) — test: fixed 3 false-failure assertions in not-logged-in.spec.js — an unscoped `iframe` locator matched Cloudflare Turnstile's injected widget on prod (strict-mode violation) in 2 tests, and a stray leading space made a title regex unmatchable in a 3rd; the title regex was corrected in place, and the 2 iframe assertions were replaced with Turnstile-safe body-text error-message checks (matching the pattern already used elsewhere in the file) so the success-vs-error discriminator isn't lost. Follow-up tracked in [#1648](https://github.com/elan-registry/registry/issues/1648): no test currently asserts the pdf-viewer.php iframe positively renders the correct document `src`, deferred until the Cloudflare Turnstile interference can be worked around
 - [#1538](https://github.com/elan-registry/registry/issues/1538) — fix: pdf-viewer.php wrapper pages flagged as Soft 404 by Google — added a hand-authored per-document metadata map (unique `$pageTitle`/`$pageDescription`/on-page description text per PDF, hardcoded literal strings only, resolved before `?doc=`/`subdir` validation to satisfy the project's before-`init.php` timing convention) plus a crawlable direct download link; unmapped documents (e.g. `docs/stories/*` PDFs) fall back to generic text rather than 404ing; added the 9 reference PDF asset URLs to `sitemap.xml`; added `PdfViewerDocumentMetadataConsistencyTest` to keep the metadata map and sitemap entries from drifting apart
+- [#1283](https://github.com/elan-registry/registry/issues/1283) — refactor: extracted a shared `TransferIntegrationTestCase` base class for car-transfer integration tests — consolidated 4 independently-duplicated `createTransferRequest()`/`insertTransferRequest()` fixture helpers (across `CarTransferRepositoryIntegrationTest`, `TransferRequestTest`, `CarTransferWorkflowTest`, and `UserDeletionReassignmentTest`) into one canonical defaults array and teardown loop, and routed every insert through `CarTransferRepository::create()` instead of hand-written raw `INSERT`s; `TransferRequestConstraintTest` still has its own independent copy of the pattern (left out of scope) but a stale docblock reference to it was corrected
